MIP-100: refactor
